网格修复

+++ src/main/java/io/github/sefiraat/networks/network/stackcaches/BarrelIdentity.java
@@ -10,11 +10,11 @@ import javax.annotation.ParametersAreNonnullByDefault;
 public abstract class BarrelIdentity extends ItemStackCache implements BarrelCore {
 
     private final Location location;
-    private final int amount;
+    private final long amount;
     private final BarrelType type;
 
     @ParametersAreNonnullByDefault
-    protected BarrelIdentity(Location location, ItemStack itemStack, int amount, BarrelType type) {
+    protected BarrelIdentity(Location location, ItemStack itemStack, long amount, BarrelType type) {
         super(itemStack);
         this.location = location;
         this.amount = amount;
@@ -25,7 +25,7 @@ public abstract class BarrelIdentity extends ItemStackCache implements BarrelCor
         return this.location;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return this.amount;
     }
